Split registration between preint and block event
